+++ src/test/java/org/realityforge/gwt/symbolmap/SymbolEntryTest.java
@@ -37,7 +37,7 @@ public class SymbolEntryTest
     assertEquals( symbolEntry1.getClassName(), "arez.ArezContext" );
     assertEquals( symbolEntry1.getMemberName(), "" );
     assertEquals( symbolEntry1.getSourceUri(), "arez/ArezContext.java" );
-    assertEquals( symbolEntry1.getSourceLine(), "29" );
+    assertEquals( symbolEntry1.getSourceLine(), 29 );
     assertEquals( symbolEntry1.getFragmentNumber(), -1 );
     assertEquals( symbolEntry1.toString(), "ArezContext,,arez.ArezContext,,arez/ArezContext.java,29,-1" );
 
@@ -47,7 +47,7 @@ public class SymbolEntryTest
     assertEquals( symbolEntry2.getClassName(), "arez.ArezContext" );
     assertEquals( symbolEntry2.getMemberName(), "$action" );
     assertEquals( symbolEntry2.getSourceUri(), "arez/ArezContext.java" );
-    assertEquals( symbolEntry2.getSourceLine(), "1533" );
+    assertEquals( symbolEntry2.getSourceLine(), 1533 );
     assertEquals( symbolEntry2.getFragmentNumber(), 0 );
     assertEquals( symbolEntry2.toString(),
                   "r,arez.ArezContext::$action(Larez/ArezContext;Ljava/lang/String;Larez/TransactionMode;Larez/Procedure;ZLarez/Observer;[Ljava/lang/Object;)V,arez.ArezContext,$action,arez/ArezContext.java,1533,0" );
@@ -77,7 +77,7 @@ public class SymbolEntryTest
     assertEquals( symbolEntry1.getClassName(), "arez.ArezContext" );
     assertEquals( symbolEntry1.getMemberName(), "" );
     assertEquals( symbolEntry1.getSourceUri(), "arez/ArezContext.java" );
-    assertEquals( symbolEntry1.getSourceLine(), "29" );
+    assertEquals( symbolEntry1.getSourceLine(), 29 );
     assertEquals( symbolEntry1.getFragmentNumber(), -1 );
   }