+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Convert `SymbolEntry.fragmentNumber` to an integer.
+* Convert `SymbolEntry.sourceLine` to an integer.
 
 ### [v0.04](https://github.com/realityforge/gwt-symbolmap/tree/v0.04) (2018-06-05)
 [Full Changelog](https://github.com/realityforge/gwt-symbolmap/compare/v0.03...v0.04)